move Records/Providers > components/Summary
